requirements.txt
streamlit
langchain
openai
tiktoken
PyPDF2
Chromadb
python-docx